ai-service/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
python-dotenv==1.0.1
Pillow==10.3.0

# supabase 최신 버전(2.x) – httpx 0.27이랑 잘 맞음
supabase==2.25.0

# 🔥 여기서 httpx 버전을 확실하게 고정
httpx==0.27.2

# OpenAI SDK – 지금 main.py에서 쓰는 버전
openai==1.3.5

torch==2.4.0
diffusers==0.30.0
transformers==4.41.1
accelerate==0.30.0
sentencepiece==0.1.99
protobuf==4.25.3
